Refactor index route and App.
