+++ src/components/App.jsx
@@ -14,20 +14,14 @@ import createHistory from 'history/createBrowserHistory';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import Navigation from 'components/ui/shared/Navigation';
 
+import RouteWithSubRoutes from 'helpers/routes/RouteWithSubRoutes';
+
 import DevTools from 'components/containers/DevTools';
 
 import store from 'store';
 import prepareData from 'helpers/prepareData';
 
 const history = createHistory();
-
-const RouteWithSubRoutes = (route) => (
-  <Route path={route.path} render={props => (
-    route.render
-    ? route.render(props)
-    : <route.component {...props} routes={route.routes}/>
-  )}/>
-);
 
 class App extends React.Component {
   render() {
@@ -39,14 +33,14 @@ class App extends React.Component {
           <div>
             <Router history={history} >
               <div>
-            <Navigation />
-              <Switch>
-                {
-                  routes.map((route, i) => (
-                    <RouteWithSubRoutes key={i} {...route} />
-                  ))
-                }
-              </Switch>
+                <Navigation />
+                <Switch>
+                  {
+                    routes.map((route, i) => (
+                      <RouteWithSubRoutes key={i} {...route} />
+                    ))
+                  }
+                </Switch>
               </div>
             </Router>
           </div>
